+++ src/models/mod.rs
@@ -1,4 +1,4 @@
+pub mod project;
 pub mod session;
 pub mod target;
 pub mod user;
-pub mod workspace;